Refactoring transformation pipeline
Removed obsolete sample states (margin/empty/etc)
Removed unused processing steps (paa/random sampling)
New sample state introduced, it's called REGULLAR and it's added
to the sample when using group-aggregate query. The transformation
steps can detect regullar/irregullar series by watching only at
one sample at a time.
