+++ akumulid/query_results_pooler.cpp
@@ -503,14 +503,12 @@ std::tuple<size_t, bool> QueryResultsPooler::read_some(char *buf, size_t buf_siz
     char* end = begin + buf_size;
     while(rdbuf_pos_ < rdbuf_top_) {
         const aku_Sample* sample = reinterpret_cast<const aku_Sample*>(rdbuf_.data() + rdbuf_pos_);
-        if (sample->payload.type != aku_PData::EMPTY) {
-            char* next = formatter_->format(begin, end, *sample);
-            if (next == nullptr) {
-                // done
-                break;
-            }
-            begin = next;
+        char* next = formatter_->format(begin, end, *sample);
+        if (next == nullptr) {
+            // done
+            break;
         }
+        begin = next;
         assert(sample->payload.size);
         rdbuf_pos_ += sample->payload.size;
     }
