+++ libakumuli/query_processing/filterbyid.h
@@ -24,9 +24,6 @@ struct FilterByIdNode : std::enable_shared_from_this<FilterByIdNode<Predicate>>,
     virtual void complete() { next_->complete(); }
 
     virtual bool put(const aku_Sample& sample) {
-        if (sample.payload.type > aku_PData::MARGIN) {
-            return next_->put(sample);
-        }
         return op_(sample.paramid) ? next_->put(sample) : true;
     }
